Heart Disease Research Part II CODECADEMY
This is the complete code that i used for the codecademy project.
I was not bale to find the correct link to get help on this project as no one has uploaded it in the codecademy forums but i have made and uploaded this project in the hope that someone who needs it gets it on time and save a lott of effort by understanding these topics.